Enhance trending page, add setlists tab, improve activity tracking
